fixed show run time.
