+++ clean_data/utils/extract_values.py
@@ -42,3 +42,58 @@ def extract_date_from_opening_date(s):
     # pd.Series(new_values)
     new_values = pd.to_datetime(new_values, errors="coerce", cache=True)
     return new_values
+
+
+# ------------------------------------------------------------------------------
+
+# Clean up Running Time
+# Before doing this, need to investigate shows with multiple parts...
+def extract_time_from_running_time(x):
+    """extracts the number of minutes from the running time"""
+
+
+    # These values are null
+    if not x or type(x)!=str:
+        return None
+
+    # This is the case for one show... it's 2 hours...
+    if x=="0200":
+        return 120
+
+    pattern_dict = {
+        "one":"1",
+        "two":"2",
+        "three":"3",
+        "four":"4",
+        "five":"5",
+        "hrs":"hours",
+        "mins":"minutes",
+        }
+    pattern_dict = {re.compile(k,re.I | re.MULTILINE):v for k,v in pattern_dict.items()}
+
+    # In the string `x` – replace any keys with the values
+    for k,v in pattern_dict.items():
+        if k.search(x):
+            x = k.sub(v, x, 2)
+
+    # Instantiate minutes
+    total_n_minutes = 0
+
+    # If there are hours
+    n_hours = re.search("([0-9]+) hour", x.lower(), re.I)
+    if n_hours:
+        n_hours = n_hours.group(1)
+
+        # convert to an int and add
+        n_hours = int(n_hours)
+        total_n_minutes += 60 * n_hours
+
+    # If there are minutes
+    n_minutes = re.search("([0-9]+) minutes", x, re.I)
+    if n_minutes:
+        n_minutes = n_minutes.group(1)
+
+        # convert to an int and add
+        total_n_minutes += int(n_minutes)
+
+    return total_n_minutes
